rename rremodule
