+++ scripts/compare_blue.py
@@ -1,4 +1,4 @@
-from src.game_tools.objects import objects
+from src.game_object.objects import objects
 from src.image_tools.compare_methods import numpy_diff
 
 import numpy as np
@@ -6,9 +6,6 @@ from PIL import Image
 
 book = objects["book"]
 
-temp = next(book.template.images)
-scr = np.array(Image.open('screen.png'))
-crp = book.template.crop_screen()
 
-result = numpy_diff()
+result = book.quiq_compare()
 print(result)
